Added Close Button to LogScreen, Refactored
